requirements.txt
asyncio~=3.4.3
python-dotenv~=1.0.1
python-binance==1.0.16
binance==5.1.0
python-telegram-bot==20.1
requests==2.28.1
cryptography